Move TerminalControl and WebControl to `io`
