+++ examples/snake/src/main.rs
@@ -2,7 +2,7 @@ mod game;
 use hewn::game_object::Coordinate;
 
 #[cfg(not(target_arch = "wasm32"))]
-use hewn::{control::TerminalControl, display::BaseDisplay, initialize_terminal, TerminalRenderer};
+use hewn::{display::BaseDisplay, initialize_terminal, io::TerminalControl, TerminalRenderer};
 fn main() {
     play_snake_in_terminal();
 }
