+++ src/engine.rs
@@ -1,7 +1,6 @@
 //! TODO: Add module documentation for engine
 //! Core game engine providing rendering, input handling, and game object management
 
-pub mod control;
 pub mod display;
 pub mod game;
 pub mod game_object;
