+++ src/lib.rs
@@ -26,11 +26,12 @@ use wasm_bindgen::prelude::*;
 
 mod engine;
 
-pub use engine::control;
 pub use engine::cursor;
 pub use engine::display;
 pub use engine::game;
 pub use engine::game_object;
+pub use engine::io;
+
 #[cfg(not(target_arch = "wasm32"))]
 pub use engine::initialize_terminal;
 pub use engine::BaseDisplay;
